Skip port_refcounts check for nat44_ed static mapping entry add
The endpoint dependent mode can reuse same port of an address as long as the
session entry is unique. So the change skips refcount check for ed mode.
This change is needed to make nat44 static mapping work in ed mode when the
port is already in use by another dynamic translation